implementations/webidl-1/tests/legacy-caller-001.html

<!DOCTYPE html>
<title>Legacy callers</title>
<script src="https://w3c-test.org/resources/testharness.js"></script>
<script src="https://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.7.5 Platform object [[Call]] method
//
// Objects that implement an interface that has a legacy caller have an
// internal [[Call]] method that invokes the caller behavior.

// feature: HTMLObjectElement (interface) with legacy caller
// feature: HTMLAllCollection (interface) with legacy caller

test(function() {
  var object = document.createElement("object");
  assert_equals(typeof object, "function");
}, "typeof HTMLObjectElement is function");

test(function() {
  var allCollection = document.all;
  var elem = allCollection(2);
  assert_true(elem instanceof HTMLTitleElement);
}, "an HTMLAllCollection can be called");
</script>
